+++ packages/jest-cli/src/cli/processArgs.js
@@ -195,6 +195,35 @@ function processArgs() {
         ),
         type: 'boolean',
       },
+      testcheckTimes: {
+        default: 100,
+        description: _wrapDesc(
+          'The number of test cases to generate for each testcheck test. ' +
+          'May be overriden for individual test cases using the options ' +
+          'argument of check.it.'
+        ),
+        type: 'number',
+      },
+      testcheckMaxSize: {
+        default: 200,
+        description: _wrapDesc(
+          'The maximum size of sized types, such as arrays and ints, to be ' +
+          'generated for testcheck tests. ' +
+          'May be overriden for individual test cases using the options ' +
+          'argument of check.it. ' +
+          'Generators can also be sized using gen.resize(n, anotherGenerator).'
+        ),
+        type: 'number',
+      },
+      testcheckSeed: {
+        default: undefined,
+        description: _wrapDesc(
+          'The seed for generating testcheck test cases. Defaults to random. ' +
+          'May be overriden for individual test cases using the options ' +
+          'argument of check.it.'
+        ),
+        type: 'number',
+      },
     })
     .check(argv => {
       if (argv.runInBand && argv.hasOwnProperty('maxWorkers')) {