+++ packages/jest-jasmine2/src/index.js
@@ -87,6 +87,11 @@ function jasmine2(config, environment, moduleLoader, testPath) {
 
     jasminePit.install(environment.global);
 
+    // Jasmine-check must be installed from within the vm context because
+    // testcheck is compiled Clojurescript and is very delicate.
+    environment.global.jasmine.testcheckOptions = config.testcheckOptions;
+    moduleLoader.requireModule(require.resolve('./jasmine-check-install'));
+
     if (config.setupTestFrameworkScriptFile) {
       moduleLoader.requireModule(config.setupTestFrameworkScriptFile);
     }