+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Windows stability fixes.
 * Mock module resolution fixes.
 * Remove test files from code coverage.
+* Add support for property testing via testcheck-js.
 
 ## jest-cli 12.1.0